Cache the implicit VariantResolvingArtifactSet for a given node at the build-tree level
We previously created an instance of these for each node in the graph that contributes artifacts to the result.
For large builds, that is a lot of nodes, and this is the greatest contributor to memory usage of a resolved Configuration.

We cache the 'implicit' artifact set for a given node, which is the artifact set that is not modified by a dependency.
These implicit artifact sets are by far the most common, as most dependencies do not modify artifacts (they do not have attributes, capabilities, excludes, or requested artifacts)

This should result in a relatively significant memory usage improvement