Optimize artifact selection
Optimize artifact selection after reviewing flamegraphs.
We pull up a cache used in ConsumerProvidedVariantFinder that simply wrapped an AttributeMatcher. Now that we
are de-duplicating these attribute matchers, we can pull up the cache into the matcher. This also allows us to
use the cached values for cases where we were previously calling the wrapped method directly on the matcher and
not through the cache.

Furthermore, we try to again reduce contention on the immutable attributes factory. This remains a hot spot
in the flame graphs when concating attributes from the consumer-side artifact type registry to the producer-side
variant artifact sets